requirements.txt
torch==2.2.2
torchvision==0.17.2
transformers==4.44.2
pillow